refactor(accounts): extract site name helper
